Revert "Merge branch 'master' into remember-playing-queue/245"
This reverts commit 76f5cf4d7c3252679808c0282142080017f6b2c7, reversing
changes made to 338ca38d0d133f23d84cc6a12a3926eab0210190.

seems to remove quite a lot of files that break the pipeline, waiting
untill comments are adressed

+++ server/subsonic/browsing.go
@@ -354,8 +354,8 @@ func (api *Router) buildArtistDirectory(ctx context.Context, artist *model.Artis
 	if artist.PlayCount > 0 {
 		dir.Played = &artist.PlayDate
 	}
-	dir.AlbumCount = int32(artist.AlbumCount)
-	dir.UserRating = int32(artist.Rating)
+	dir.AlbumCount = artist.AlbumCount
+	dir.UserRating = artist.Rating
 	if artist.Starred {
 		dir.Starred = &artist.StarredAt
 	}
@@ -392,8 +392,8 @@ func (api *Router) buildAlbumDirectory(ctx context.Context, album *model.Album) 
 	if album.PlayCount > 0 {
 		dir.Played = &album.PlayDate
 	}
-	dir.UserRating = int32(album.Rating)
-	dir.SongCount = int32(album.SongCount)
+	dir.UserRating = album.Rating
+	dir.SongCount = album.SongCount
 	dir.CoverArt = album.CoverArtID().String()
 	if album.Starred {
 		dir.Starred = &album.StarredAt
@@ -415,15 +415,15 @@ func (api *Router) buildAlbum(ctx context.Context, album *model.Album, mfs model
 	dir.Artist = album.AlbumArtist
 	dir.ArtistId = album.AlbumArtistID
 	dir.CoverArt = album.CoverArtID().String()
-	dir.SongCount = int32(album.SongCount)
-	dir.Duration = int32(album.Duration)
+	dir.SongCount = album.SongCount
+	dir.Duration = int(album.Duration)
 	dir.PlayCount = album.PlayCount
 	if album.PlayCount > 0 {
 		dir.Played = &album.PlayDate
 	}
-	dir.Year = int32(album.MaxYear)
+	dir.Year = album.MaxYear
 	dir.Genre = album.Genre
-	dir.UserRating = int32(album.Rating)
+	dir.UserRating = album.Rating
 	if !album.CreatedAt.IsZero() {
 		dir.Created = &album.CreatedAt
 	}
