+++ server/subsonic/searching.go
@@ -104,8 +104,8 @@ func (api *Router) Search2(r *http.Request) (*responses.Subsonic, error) {
 		searchResult2.Artist[i] = responses.Artist{
 			Id:             artist.ID,
 			Name:           artist.Name,
-			AlbumCount:     int32(artist.AlbumCount),
-			UserRating:     int32(artist.Rating),
+			AlbumCount:     artist.AlbumCount,
+			UserRating:     artist.Rating,
 			CoverArt:       artist.CoverArtID().String(),
 			ArtistImageUrl: public.ImageURL(r, artist.CoverArtID(), 600),
 		}
